==> src/step01_tsv2anndata.R <==


library(reticulate)
use_python('/home/toolkit/local/bin/python3',required=TRUE)

########################
library(pROC)
library(MuSiC)
library(Biobase)
library(Seurat)
library(SeuratData)
library(SeuratDisk)
library(anndata)

source('fwp.R')
source('fwp_source.R')

source('fwp_source.R')
this_path='./data/first/'
TYPE=c('BRCA','KIRP','LIHC','PCPG')
MODE=c('train','test')

USED_GENE=.getUsedGene(this_path,TYPE,MODE)

set.seed(123)
i=1
while(i<=length(TYPE)){
j=1
while(j<=length(MODE)){
#####################################
this_type=TYPE[i]
this_mode=MODE[j]
print(this_type)
print(this_mode)
.preAnn(this_path, this_type, this_mode, USED_GENE, CUT=10)
#####################################
j=j+1}
i=i+1}


source('fwp_source.R')
this_path='./data/second/'
TYPE=c('BLCA','CESC','CHOL','COAD','ESCA','HNSC','KICH','LUSC','PAAD','THYM','UCEC')
MODE=c('train','test')

USED_GENE=.getUsedGene(this_path,TYPE,MODE)

set.seed(123)
i=1
while(i<=length(TYPE)){
j=1
while(j<=length(MODE)){
#####################################
this_type=TYPE[i]
this_mode=MODE[j]
print(this_type)
print(this_mode)
.preAnn(this_path, this_type, this_mode, USED_GENE, CUT=10)
#####################################
j=j+1}
i=i+1}
